pyproject.toml
[tool.poetry]
name = "karaoke-decide"
version = "0.1.0"
description = "Help people discover and choose the perfect karaoke songs based on their music listening history"
authors = ["Andrew Beveridge <andrew@beveridge.uk>"]
readme = "README.md"
packages = [
    { include = "karaoke_decide" },
    { include = "backend" },
]

[tool.poetry.scripts]
karaoke-decide = "karaoke_decide.cli.main:main"

[tool.poetry.dependencies]
python = "^3.11"
# CLI
click = "^8.1.0"
rich = "^13.7.0"
# API Framework
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }
# Data validation
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
# Google Cloud
google-cloud-firestore = "^2.14.0"
google-cloud-storage = "^2.14.0"
google-cloud-secret-manager = "^2.16.0"
# HTTP clients
httpx = "^0.27.0"
aiohttp = "^3.9.0"
# Auth
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
# Email
sendgrid = "^6.11.0"
# Utilities
python-slugify = "^8.0.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
pytest-mock = "^3.12.0"
ruff = "^0.8.0"
mypy = "^1.8.0"
pre-commit = "^4.0.0"
respx = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/integration", "backend/tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--cov=karaoke_decide",
    "--cov=backend",
    "--cov-report=term-missing",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration",
    "emulator: marks tests requiring GCP emulators",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::RuntimeWarning",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true
